pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "blauw-zwart-fan-sim-pipeline"
version = "0.1.0"
requires-python = ">=3.12"
# IANA time zone database for zoneinfo (required on Windows for Europe/Brussels etc.).
dependencies = [
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openrouter>=0.2.1",
    "langgraph>=1.1.9",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.2",
    "sqlglot[c]>=30.6.0",
    "tzdata>=2024.1",
]

[project.optional-dependencies]
# Kafka output for `fan_events stream --kafka-topic`; install with: uv sync --extra kafka
kafka = ["confluent-kafka>=2.3"]
ingest = [
    "asyncpg>=0.29",
    "confluent-kafka>=2.3",
]
# Natural-language Q&A API backed by Ollama + Postgres marts: `uv sync --extra api`
api = [
    "flask>=3.0",
    "psycopg2-binary>=2.9",
    "requests>=2.31",
    "pyyaml>=6.0",
]
# Pro League scraper service: `uv sync --extra scraper`
scraper = [
    "beautifulsoup4>=4.12",
    "confluent-kafka>=2.3",
    "flask>=3.0",
    "lxml>=5.0",
    "psycopg2-binary>=2.9",
    "requests>=2.31",
]
# Pro League Kafka→Postgres ingest consumer: `uv sync --extra proleague-ingest`
proleague-ingest = [
    "confluent-kafka>=2.3",
    "psycopg2-binary>=2.9",
]

[project.scripts]
fan_events = "fan_events.cli:main"
fan_ingest = "fan_ingest.main:main"

[dependency-groups]
dev = [
    "asyncpg>=0.29",
    "beautifulsoup4>=4.14.3",
    "confluent-kafka>=2.14.0",
    "flask>=3.0",
    "psycopg2-binary>=2.9",
    "pyyaml>=6.0",
    "requests>=2.31",
    "ipykernel>=7.2.0",
    "lxml>=6.0.4",
    "pytest>=9.0.2",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
]
# dbt against local Compose Postgres: `uv sync --group dbt` then `uv run dbt ...`.
# On Windows, prefer `uv run dbt` over a global `dbt` if that binary is dbt-fusion (Postgres DLL errors).
dbt = [
    "dbt-core~=1.8",
    "dbt-postgres~=1.8",
]

[tool.coverage.run]
branch = true
source = ["src"]
omit = [
    "*/__init__.py",
    "*/__main__.py",
    "src/blauw_zwart_fan_sim_pipeline.egg-info/*",
]

[tool.coverage.report]
show_missing = true
skip_empty = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
# Resolved from the repo root; run `uv run pytest` there (or pass `../tests` if invoking from `src/`).
testpaths = ["tests"]
markers = [
    "slow: optional large-batch or long-running tests (e.g. SC-003 retail scale)",
]

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [".agents", ".cursor", ".git", ".specify", ".venv", "venv"]

[tool.ruff.format]

[tool.ruff.lint]
exclude = ["*.ipynb"]
select = ["E", "F", "I", "W"]
